java: fix various integer-math error-prone issues
- Adds extra () around various bitwise math to clarify operator
  precedence (non-functional)

- Adds an upcast to 'long' to avoid potential overflow calculating the
  size of a write request. This shouldn't overflow in practice given
  other limits on write batch sizes.

- Removed some unused methods in Slices.java

This should be non-functional.

Reviewed-on: http://gerrit.cloudera.org:8080/10200
Tested-by: Kudu Jenkins

+++ java/kudu-client/src/main/java/org/apache/kudu/client/Batch.java
@@ -91,8 +91,8 @@ class Batch extends KuduRpc<BatchResponse> {
   Message createRequestPB() {
     final Tserver.WriteRequestPB.Builder builder =
         Operation.createAndFillWriteRequestPB(operations);
-    rowOperationsSizeBytes = builder.getRowOperations().getRows().size() +
-                             builder.getRowOperations().getIndirectData().size();
+    rowOperationsSizeBytes = (long)builder.getRowOperations().getRows().size() +
+                             (long)builder.getRowOperations().getIndirectData().size();
     builder.setTabletId(UnsafeByteOperations.unsafeWrap(getTablet().getTabletIdAsBytes()));
     builder.setExternalConsistencyMode(externalConsistencyMode.pbVersion());
     return builder.build();
