+++ java/kudu-client/src/main/java/org/apache/kudu/client/Operation.java
@@ -130,8 +130,8 @@ public abstract class Operation extends KuduRpc<OperationResponse> {
   Message createRequestPB() {
     final Tserver.WriteRequestPB.Builder builder =
         createAndFillWriteRequestPB(ImmutableList.of(this));
-    this.rowOperationSizeBytes = builder.getRowOperations().getRows().size() +
-        builder.getRowOperations().getIndirectData().size();
+    this.rowOperationSizeBytes = (long)builder.getRowOperations().getRows().size() +
+        (long)builder.getRowOperations().getIndirectData().size();
     builder.setTabletId(UnsafeByteOperations.unsafeWrap(getTablet().getTabletIdAsBytes()));
     builder.setExternalConsistencyMode(this.externalConsistencyMode.pbVersion());
     if (this.propagatedTimestamp != AsyncKuduClient.NO_TIMESTAMP) {
